server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
psycopg[binary]>=3.2.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.2.0
email-validator>=2.2.0
httpx>=0.27.0
python-dotenv>=1.0.0
